again edited data struct
byte-float mistakes

+++ GS_receiver_00/Data_Structs.h
@@ -1,17 +1,17 @@
 struct Data_Package {
   char state;
-  double temperature;
-  double pressure;
-  byte Alt;
-  byte Roll;
-  byte Pitch;
-  byte Yaw;
+  float temperature;
+  float pressure;
+  float Alt;
+  float Roll;
+  float Pitch;
+  float Yaw;
   byte SCount;
   uint32_t gps_t;
   float gps_alt;
   float Lat;
   float Long;
-  byte Speed;
+  float Speed;
   float Voltage;
 
 };
